refactor: PR レビュー指摘 (round 2) を反映
- myPlayerAtom を MyPlayerProvider/useMyPlayerValue に置き換え。
  useQuery の戻り値と atom の 1 render ずれを構造的に解消し、Context
  経路に統一。app スコープを意識する必要は実利用上少なく、ページ間
  での再 fetch コストも問題にならない。
- useIconsValue を Provider 外で例外を投げる形に変更し、他の Context
  hooks（useGameValue / useMyselfValue / useTalkPanel 等）と一貫させる。
